Move custom serialization to `PlatformExtensionConverter`s
- The gist is that we can utilize generic JsonConverters in order to deserialize portions of client capabilities when interacting with clients. That way sub-systems can try and cast those client capabilities to corresponding types and succeed with minimal code requirements.
- This is setting the groundwork for doing a lot more custom extension serialization in our language server.
- Updated tests to also accoutn for the changes.

Part of dotnet/aspnetcore#31883
Part of dotnet/aspnetcore#31886
